requirements.txt
attrs==21.2.0
click==8.0.3
Flask==2.0.2
flask-expects-json==1.7.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
importlib-metadata==4.8.2
itsdangerous==2.0.1
Jinja2==3.0.3
jsonschema==3.2.0
MarkupSafe==2.0.1
PyJWT==2.3.0
pyrsistent==0.18.0
six==1.16.0
SQLAlchemy==1.4.27
typing-extensions==4.0.0
Werkzeug==2.0.2
zipp==3.6.0
gunicorn==20.1.0
